backend/django/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
openai==1.3.0
python-dotenv==1.0.0
